requirements.txt
streamlit 
langchain 
chromadb 
sentence-transformers 
pymupdf 
python-dotenv 
tiktoken 
ollama
pdfplumber
langchain-community
pandas
openpyxl